+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+-  Added agent language as filter to avoid displaying incompatible modules
+-  Added folder display to agent modules
+
 ## [3.0.0] - 2025-03-25
 
 ### Added